# cmd/apps/apis/rest/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  cmd_apps_apis_rest_v1.ListWorkflowJobsResponse:
    properties:
      jobs:
        items:
          $ref: '#/definitions/cmd_apps_apis_rest_v1.WorkflowJobDetail'
        type: array
    type: object
  cmd_apps_apis_rest_v1.ListWorkflowsResponse:
    properties:
      workflows:
        items:
          $ref: '#/definitions/cmd_apps_apis_rest_v1.WorkflowInfo'
        type: array
    type: object
...
  cmd_apps_apis_rest_v1.TriggerWorkflowRequest:
    properties:
      reason:
        type: string
      timeout:
        type: integer
      uri:
        type: string
    type: object
  cmd_apps_apis_rest_v1.TriggerWorkflowResponse:
    properties:
      job_id:
        type: string
    type: object
...
  cmd_apps_apis_rest_v1.WorkflowInfo:
    properties:
      created_at:
        type: string
      id:
        type: string
      last_triggered_at:
        type: string
      name:
        type: string
      queue_name:
        type: string
      updated_at:
        type: string
    type: object
  cmd_apps_apis_rest_v1.WorkflowJobDetail:
    properties:
      created_at:
        type: string
      finish_at:
        type: string
      id:
        type: string
      message:
        type: string
      queue_name:
        type: string
      start_at:
        type: string
      status:
        type: string
      steps:
        items:
          $ref: '#/definitions/v1.WorkflowJobStep'
        type: array
      target:
        $ref: '#/definitions/v1.WorkflowJobTarget'
      trigger_reason:
        type: string
      updated_at:
        type: string
      workflow:
        type: string
    type: object
...
  types.WorkflowLocalFileWatch:
    properties:
      cel_pattern:
        type: string
      directory:
        type: string
      event:
        type: string
      file_pattern:
        type: string
      file_types:
        type: string
      max_file_size:
        type: integer
      min_file_size:
        type: integer
    type: object
  types.WorkflowNode:
    properties:
      branches:
        additionalProperties:
          type: string
        type: object
      cases:
        description: switch
        items:
          $ref: '#/definitions/types.WorkflowNodeCase'
        type: array
      condition:
        description: if
        type: string
      default:
        type: string
      input:
        additionalProperties: {}
        type: object
      matrix:
        allOf:
        - $ref: '#/definitions/types.WorkflowNodeMatrix'
        description: loop
      name:
        type: string
      next:
        type: string
      params:
        additionalProperties:
          type: string
        type: object
      type:
        type: string
    type: object
  types.WorkflowNodeCase:
    properties:
      next:
        type: string
      value:
        type: string
    type: object
  types.WorkflowNodeMatrix:
    properties:
      data:
        additionalProperties: {}
        description: |-
          Data defines variable mappings for matrix iteration
          Key is variable name, value is template reference to context data
          Example: {"file_path": "{{ new_webpage.file_paths }}"}
        type: object
    type: object
  types.WorkflowRssTrigger:
    type: object
  types.WorkflowTrigger:
    properties:
      interval:
        type: integer
      local_file_watch:
        $ref: '#/definitions/types.WorkflowLocalFileWatch'
      rss:
        $ref: '#/definitions/types.WorkflowRssTrigger'
    type: object
...
  v1.CreateWorkflowRequest:
    properties:
      enable:
        type: boolean
      name:
        type: string
      nodes:
        items:
          $ref: '#/definitions/types.WorkflowNode'
        type: array
      queue_name:
        type: string
      trigger:
        $ref: '#/definitions/types.WorkflowTrigger'
    required:
    - name
    type: object
  v1.CreateWorkflowResponse:
    properties:
      workflow:
        $ref: '#/definitions/cmd_apps_apis_rest_v1.WorkflowInfo'
    type: object
...
  v1.MessageResponse:
    properties:
      message:
        type: string
    type: object
...
  v1.UpdateWorkflowRequest:
    properties:
      enable:
        type: boolean
      name:
        type: string
      nodes:
        items:
          $ref: '#/definitions/types.WorkflowNode'
        type: array
      queue_name:
        type: string
      trigger:
        $ref: '#/definitions/types.WorkflowTrigger'
    type: object
  v1.WorkflowJobResponse:
    properties:
      job:
        $ref: '#/definitions/cmd_apps_apis_rest_v1.WorkflowJobDetail'
    type: object
  v1.WorkflowJobStep:
    properties:
      message:
        type: string
      name:
        type: string
      status:
        type: string
    type: object
  v1.WorkflowJobTarget:
    properties:
      entries:
        items:
          type: string
        type: array
    type: object
  v1.WorkflowResponse:
    properties:
      workflow:
        $ref: '#/definitions/cmd_apps_apis_rest_v1.WorkflowInfo'
    type: object
host: localhost:7081
info:
  contact: {}
  description: NanaFS is a Reference Filing System API
  title: NanaFS API
  version: "1.0"
paths:
...
  /api/v1/workflows:
    get:
      consumes:
      - application/json
      description: Retrieve all available workflows
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/cmd_apps_apis_rest_v1.ListWorkflowsResponse'
      summary: List workflows
      tags:
      - Workflows
    post:
      consumes:
      - application/json
      description: Create a new workflow
      parameters:
      - description: Create workflow request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/v1.CreateWorkflowRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/v1.CreateWorkflowResponse'
      summary: Create workflow
      tags:
      - Workflows
  /api/v1/workflows/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a workflow by ID
      parameters:
      - description: Workflow ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v1.MessageResponse'
      summary: Delete workflow
      tags:
      - Workflows
    get:
      consumes:
      - application/json
      description: Retrieve a specific workflow by ID
      parameters:
      - description: Workflow ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v1.WorkflowResponse'
      summary: Get workflow
      tags:
      - Workflows
    put:
      consumes:
      - application/json
      description: Update an existing workflow
      parameters:
      - description: Workflow ID
        in: path
        name: id
        required: true
        type: string
      - description: Update workflow request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/v1.UpdateWorkflowRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v1.WorkflowResponse'
      summary: Update workflow
      tags:
      - Workflows
  /api/v1/workflows/{id}/jobs:
    get:
      consumes:
      - application/json
      description: Retrieve all jobs for a specific workflow
      parameters:
      - description: Workflow ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/cmd_apps_apis_rest_v1.ListWorkflowJobsResponse'
      summary: List workflow jobs
      tags:
      - Workflows
  /api/v1/workflows/{id}/jobs/{jobId}:
    get:
      consumes:
      - application/json
      description: Retrieve a specific job by workflow ID and job ID
      parameters:
      - description: Workflow ID
        in: path
        name: id
        required: true
        type: string
      - description: Job ID
        in: path
        name: jobId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v1.WorkflowJobResponse'
      summary: Get workflow job
      tags:
      - Workflows
  /api/v1/workflows/{id}/jobs/{jobId}/cancel:
    post:
      consumes:
      - application/json
      description: Cancel a workflow job
      parameters:
      - description: Workflow ID
        in: path
        name: id
        required: true
        type: string
      - description: Job ID
        in: path
        name: jobId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v1.MessageResponse'
      summary: Cancel job
      tags:
      - Workflows
  /api/v1/workflows/{id}/jobs/{jobId}/pause:
    post:
      consumes:
      - application/json
      description: Pause a running workflow job
      parameters:
      - description: Workflow ID
        in: path
        name: id
        required: true
        type: string
      - description: Job ID
        in: path
        name: jobId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v1.MessageResponse'
      summary: Pause job
      tags:
      - Workflows
  /api/v1/workflows/{id}/jobs/{jobId}/resume:
    post:
      consumes:
      - application/json
      description: Resume a paused workflow job
      parameters:
      - description: Workflow ID
        in: path
        name: id
        required: true
        type: string
      - description: Job ID
        in: path
        name: jobId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v1.MessageResponse'
      summary: Resume job
      tags:
      - Workflows
  /api/v1/workflows/{id}/trigger:
    post:
      consumes:
      - application/json
      description: Manually trigger a workflow to run
      parameters:
      - description: Workflow ID
        in: path
        name: id
        required: true
        type: string
      - description: Trigger workflow request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/cmd_apps_apis_rest_v1.TriggerWorkflowRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/cmd_apps_apis_rest_v1.TriggerWorkflowResponse'
      summary: Trigger workflow
      tags:
      - Workflows
swagger: "2.0"